% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geetodf.R
\name{geetodf}
\alias{geetodf}
\title{outputs data frame in tidy format from gee list}
\usage{
geetodf(list)
}
\arguments{
\item{list}{the list that is returned from one of the xxtolist() functions}
}
\value{

}
\description{
this function will take the list generated by the xxtolist() functions and outputs into datafram with coordinates and data in tidy format
}
\examples{
ROI = ROI()
list = wttolist(ROI, "20010928", 5, 30000)
df = geetodf(list)

}
