% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{start}
\alias{start}
\title{A developer tool that should be deleted with final publish}
\usage{
start(path = "C:\\\\Users\\\\jacei\\\\anaconda3\\\\envs\\\\rgee_py")
}
\arguments{
\item{path}{}
}
\value{

}
\description{
A developer tool that should be deleted with final publish
}
