% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin.R
\name{thin}
\alias{thin}
\title{this is a function that will thin out the number of arrows in a vector plot}
\usage{
thin(df, by)
}
\arguments{
\item{by}{}
}
\value{

}
\description{
this is a function that will thin out the number of arrows in a vector plot
}
\examples{
vect = thin(df)
}
