% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wttolist.R
\name{wttolist}
\alias{wttolist}
\title{outputs list of water Temperate data with specified parameters}
\usage{
wttolist(ROI, date, interval, scale, depth)
}
\arguments{
\item{ROI}{an ee.geometry object, created by the ROI() function}

\item{date}{a string in "YYYYMMDD" format. (ultimate day of interval)}

\item{interval}{an integer to specify the number of days in the interval to average}

\item{scale}{the resolution in meters that sampling takes place (must output less than 5000 points)}

\item{depth}{the depth at which the data will be pulled (read documentation for specified depths)}
}
\value{

}
\description{
The point of this function is to extract data from a specific region (ROI), time (date), resolution (scale), avraged across a specified number of days (interval), from the HYCOM numerical model hosted on Google Earth Engine.
}
\examples{
ROI = ROI()
list = wttolist(ROI, "20010928", 5, 30000, 10)
}
